+++ app/controllers/concerns/backend_page.rb
@@ -46,7 +46,7 @@ module BackendPage
       time_format: Setting.get("time_format"),
       default_language: Setting.get("default_language"),
       default_timezone: Setting.get("default_timezone"),
-      fixed_timezone: Setting.fixed_timezone?,
+      timezone_support: Setting.timezone_support?,
       first_weekday: Setting.get("first_weekday")
     )
   end